+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to `laravel-login-links` will be documented in this file.
 
+## 0.30 - 2021-07-11
+- Added `generateLoginLink()` method for classes that have `CanLoginWithLink` trait.
+- Added tests for `generateLoginLink()`
+
 ## 0.20 - 2021-07-11
 - Added `LoginLinkGenerated` and `LoginLinkUsed` events.
 - Added tests for events (and for custom config)